Launcherを基本形のものとプロジェクト側のものとに分けた
