+++ Assets/App/Scripts/Common/Launcher.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 207623a2e1859472690b1ebf8a293780
+guid: 9f5f6e1fe2a19442590945f48e419044
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
