+++ Assets/App/Scripts/Common/MasterData.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 972874db81a7b404aa291230310fb894
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
