+++ Assets/App/Scripts/Installer.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 207623a2e1859472690b1ebf8a293780
+guid: 069ff43264e9f4f46a943bf99e505d96
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
