+++ Assets/App/Scripts/Models.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e8a6d0d93406d41689d545d3dfe77212
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
